+++ src/main/java/core/di/factory/BeanScanner.java
@@ -3,9 +3,6 @@ package core.di.factory;
 import core.annotation.Repository;
 import core.annotation.Service;
 import core.annotation.web.Controller;
-import core.annotation.web.RequestMapping;
-import core.mvc.tobe.HandlerExecution;
-import core.mvc.tobe.HandlerKey;
 import core.mvc.tobe.support.*;
 import org.reflections.Reflections;
 import org.reflections.scanners.MethodAnnotationsScanner;
@@ -17,7 +14,6 @@ import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
 import org.springframework.core.ParameterNameDiscoverer;
 
 import java.lang.annotation.Annotation;
-import java.lang.reflect.Method;
 import java.util.*;
 
 import static java.util.Arrays.asList;
@@ -26,7 +22,7 @@ public class BeanScanner {
 
     private static final Logger logger = LoggerFactory.getLogger(BeanScanner.class);
 
-    private static final List<ArgumentResolver> argumentResolvers = asList(
+    public static final List<ArgumentResolver> argumentResolvers = asList(
                 new HttpRequestArgumentResolver(),
                 new HttpResponseArgumentResolver(),
                 new RequestParamArgumentResolver(),
@@ -34,40 +30,19 @@ public class BeanScanner {
                 new ModelArgumentResolver()
         );
 
-    private static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
+    public static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
 
-    public Map<HandlerKey, HandlerExecution> scan(Object... basePackage) {
+    public static BeanFactory scan(Object... basePackage) {
         Reflections reflections = new Reflections(basePackage, new TypeAnnotationsScanner(), new SubTypesScanner(), new MethodAnnotationsScanner());
-        Map<HandlerKey, HandlerExecution> handlers = new HashMap<>();
-
         Set<Class<?>> preInstanticateBeans = getTargetAnnotationClass(reflections, Controller.class, Service.class, Repository.class);
 
         BeanFactory beanFactory = new BeanFactory(preInstanticateBeans);
         beanFactory.initialize();
 
-        Set<Class<?>> controllers = beanFactory.getControllerClass();
-        for (Class<?> controller : controllers) {
-            Object target = beanFactory.getBean(controller);
-            addHandlerExecution(handlers, target, controller.getMethods());
-        }
-
-        return handlers;
+        return beanFactory;
     }
 
-
-    private void addHandlerExecution(Map<HandlerKey, HandlerExecution> handlers, final Object target, Method[] methods) {
-        Arrays.stream(methods)
-                .filter(method -> method.isAnnotationPresent(RequestMapping.class))
-                .forEach(method -> {
-                    RequestMapping requestMapping = method.getAnnotation(RequestMapping.class);
-                    HandlerKey handlerKey = new HandlerKey(requestMapping.value(), requestMapping.method());
-                    HandlerExecution handlerExecution = new HandlerExecution(nameDiscoverer, argumentResolvers, target, method);
-                    handlers.put(handlerKey, handlerExecution);
-                    logger.info("Add - method: {}, path: {}, HandlerExecution: {}", requestMapping.method(), requestMapping.value(), method.getName());
-                });
-    }
-
-    private Set<Class<?>> getTargetAnnotationClass(Reflections reflections, Class<? extends Annotation>... annotations) {
+    private static Set<Class<?>> getTargetAnnotationClass(Reflections reflections, Class<? extends Annotation>... annotations) {
         Set<Class<?>> beans = new HashSet<>();
         for (Class<? extends Annotation> annotation : annotations) {
             beans.addAll(reflections.getTypesAnnotatedWith(annotation));
